+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - avoid showing horizontal scrollbars in chat list #4253
 - style: fix VCard color being too bright in dark theme #4255
 - remove unnecessary horizontal scrollbar in "View Group" dialog #4254
+- add missing cancel buttons to import-/export- and reveive-backup progress dialogs #4272
 
 <a id="1_47_0"></a>